# Add cuda event to enable async move