--- a/patrickstar/core/chunk_data.py
+++ b/patrickstar/core/chunk_data.py
@@ -31,6 +31,7 @@ import time
 import torch
 
 from patrickstar.core.memtracer import RuntimeMemTracer
+from patrickstar.manager.cuda_context import CUDAContext
 from patrickstar.profiler import profiler
 from patrickstar.utils import logger, getsizeof
 import patrickstar.utils.global_timer as global_timer
@@ -90,6 +91,8 @@ class Chunk(object):
         if self.with_mem_cache:
             self.memory_cache = memory_cache
 
+        self.compute_finish_event = torch.cuda.Event()
+
     def is_dummy(self):
         return self._is_dummy
 
@@ -127,19 +130,12 @@ class Chunk(object):
         if self._time_profile:
             global_timer.my_timer.start_profile(f"CHUNK_allocate_payload_{device.type}")
         # reuse the chunk in cache if possible
-        if self.with_mem_cache:
-            try:
+        try:
+            if self.with_mem_cache:
                 self.payload = self.memory_cache.pop_or_allocate(
                     device, payload_numel, self.data_type, device.type == "cpu"
                 )
-            except RuntimeError:
-                if self._time_profile:
-                    global_timer.my_timer.finish_profile(
-                        f"CHUNK_allocate_payload_{device.type}"
-                    )
-                return False
-        else:
-            try:
+            else:
                 self.payload = torch.zeros(
                     payload_numel,
                     dtype=self.data_type,
@@ -147,12 +143,12 @@ class Chunk(object):
                     pin_memory=(device.type == "cpu"),
                 )
                 self.memory_tracer.add(device.type, self.get_payload_space())
-            except RuntimeError:
-                if self._time_profile:
-                    global_timer.my_timer.finish_profile(
-                        f"CHUNK_allocate_payload_{device.type}"
-                    )
-                return False
+        except RuntimeError:
+            if self._time_profile:
+                global_timer.my_timer.finish_profile(
+                    f"CHUNK_allocate_payload_{device.type}"
+                )
+            return False
 
         if profiler.started():
             profiler.chunk_life_cycle[self.chunk_id]["life_cycle"].append(
@@ -189,6 +185,12 @@ class Chunk(object):
         """
         self._state_dict[old_state] -= 1
         self._state_dict[new_state] += 1
+        if (
+            old_state == TensorState.COMPUTE
+            and self._state_dict[TensorState.COMPUTE] == 0
+        ):
+            cuda_ctx = CUDAContext()
+            self.compute_finish_event.record(cuda_ctx.compute_stream)
 
     def get_state(self):
         """
@@ -265,15 +267,17 @@ class Chunk(object):
             f"used mem {self.memory_tracer.used_chunk_mem(target_device.type) / 1e6} MB"
         )
 
+        cuda_ctx = CUDAContext()
         if self.with_mem_cache:
             payload_numel = self.payload.numel()
-            # TODO(jiaruifang) asyc copy.
             if target_device.type == "cpu":
                 pinned_payload_cpu = self.memory_cache.pop_or_allocate(
                     target_device, payload_numel, self.payload.dtype, True
                 )
                 pinned_payload_cpu.reshape(self.payload.shape)
-                pinned_payload_cpu.copy_(self.payload)
+                self.compute_finish_event.synchronize()
+                with torch.cuda.stream(cuda_ctx.copy_stream):
+                    pinned_payload_cpu.copy_(self.payload)
                 self.memory_cache.push(self.payload)
                 self.payload = pinned_payload_cpu
             elif target_device.type == "cuda":
@@ -282,7 +286,9 @@ class Chunk(object):
                     target_device, payload_numel, self.payload.dtype, False
                 )
                 cuda_tmp_payload.reshape(self.payload.shape)
-                cuda_tmp_payload.copy_(self.payload)
+                self.compute_finish_event.synchronize()
+                with torch.cuda.stream(cuda_ctx.copy_stream):
+                    cuda_tmp_payload.copy_(self.payload)
                 self.memory_cache.push(self.payload)
                 self.payload = cuda_tmp_payload
         else:
@@ -293,11 +299,15 @@ class Chunk(object):
                     device="cpu:0",
                     pin_memory=True,
                 )
-                pinned_payload_cpu.copy_(self.payload)
+                self.compute_finish_event.synchronize()
+                with torch.cuda.stream(cuda_ctx.copy_stream):
+                    pinned_payload_cpu.copy_(self.payload)
                 self.payload = pinned_payload_cpu
             elif target_device.type == "cuda":
                 self.payload = self.payload.pin_memory()
-                self.payload = self.payload.to(target_device)
+                self.compute_finish_event.synchronize()
+                with torch.cuda.stream(cuda_ctx.copy_stream):
+                    self.payload = self.payload.to(target_device)
 
             self.memory_tracer.delete(src_device.type, self.get_payload_space())
             self.memory_tracer.add(target_device.type, self.get_payload_space())